deleted creategroup component, looks like figma
