+++ meetbd/src/newstyles.css
@@ -41,6 +41,16 @@ button {
     padding: 0.3rem 0.5rem;
 }
 
+input {
+    border-radius: 13px;
+    border-width: 2px;
+    border-style: solid;
+    border-color: #EDCCD8;
+    height: 30px;
+    padding-left: 10px;
+    padding-right: 10px;
+}
+
 .flash {
     animation-name: flash;
     animation-duration: 0.5s;
@@ -160,3 +170,24 @@ button {
             border-right: 0.2px solid grey;
         }
     
+.bottom {
+    bottom: 10px;
+    position: fixed;
+    width: 100%;
+}
+.add-group {
+    background-color: #F4F4F4;
+    border-color: #EDCCD8;
+    border-radius: 30px;
+    border-style: solid;
+    border-width: 3px;
+    margin-left: auto;
+    margin-right: auto;
+    padding: 15px;
+    text-align: center;
+    width: 80%;
+}
+
+.add-group h2 {
+    margin-top: 1rem;
+}
